+++ docs/changelog.rst
@@ -27,6 +27,7 @@ New Features
 - Now accounting for the idea that like-minded agents will push each other towards more extreme opinions even if the average opinion
   is currently stable (:meth:`~gatoh.model.ABModel.iteration_opinion_calculation`)
 - Added new checks to :meth:`~gatoh.graphs.Graph.agent_opinion_change` that account for agent deradicalisation
+- Added multiple QoL setter functions to :class:`~gatoh.model.ABModel`
 
 Fixes
 -----